refactor: move table into separate component
